requirements.txt
atomicwrites==1.4.1
attrs==22.1.0
backcall==0.2.0
certifi==2022.9.24
charset-normalizer==2.0.12
colorama==0.4.6
decorator==5.1.1
Django==2.2.16
django-filter==21.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.2
flake8==5.0.4
flake8-broken-line==0.6.0
flake8-isort==5.0.3
flake8-plugin-utils==1.3.2
flake8-return==1.2.0
idna==3.4
importlib-metadata==4.2.0
iniconfig==1.1.1
ipython==7.34.0
isort==5.10.1
jedi==0.18.2
Markdown==3.4.1
matplotlib-inline==0.1.6
mccabe==0.7.0
packaging==21.3
parso==0.8.3
pickleshare==0.7.5
pluggy==0.13.1
prompt-toolkit==3.0.33
py==1.11.0
pycodestyle==2.9.1
pyflakes==2.5.0
Pygments==2.13.0
PyJWT==2.1.0
pyparsing==3.0.9
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
pytz==2022.6
requests==2.26.0
sqlparse==0.4.3
toml==0.10.2
traitlets==5.6.0
typing-extensions==4.4.0
urllib3==1.26.13
wcwidth==0.2.5
zipp==3.11.0
